interface/requirements.txt
websocket-client
numpy
flask
flask-socketio
werkzeug
pandas
scipy
